requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.2
blinker==1.5
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
dateparser==1.1.7
docutils==0.19
exceptiongroup==1.0.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
lxml==4.9.2
outcome==1.2.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
robotframework==6.0.1
robotframework-databaselibrary==1.2.4
robotframework-pythonlibcore==4.0.0
robotframework-requests==0.9.4
robotframework-seleniumlibrary==6.0.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webdrivermanager==0.10.0
wsproto==1.2.0
zstandard==0.19.0
